documentation and changing navbars
