+++ frontend/src/app/expenses/bill-splitting/page.tsx
@@ -26,8 +26,6 @@ export default function BillSplitting() {
 
   return (
     <div className="min-h-screen bg-gray-100">
-      {/* Navigation Bar */}
-      <Navbar />
 
       <div className="px-6 pb-12">
         <div className="max-w-6xl mx-auto">
