+++ frontend/src/app/expenses/page.tsx
@@ -15,35 +15,8 @@ export default function Dashboard() {
 
   return (
     <div className="min-h-screen bg-gray-100">
-      {/* Navigation Bar */}
-      <nav className="bg-white shadow-md mb-8">
-        <div className="max-w-6xl mx-auto px-6">
-          <div className="flex justify-between items-center py-4">
-            <div className="text-2xl font-bold text-gray-800">💰 ExpenseTracker</div>
-            <div className="flex gap-2">
-              <button
-                className="px-6 py-2 rounded-lg font-medium transition bg-blue-600 text-white"
-              >
-                Dashboard
-              </button>
-              <button
-                onClick={() => router.push('/expenses/bill-splitting')}
-                className="px-6 py-2 rounded-lg font-medium transition bg-gray-100 text-gray-700 hover:bg-gray-200"
-              >
-                Bill Splitting
-              </button>
-              <button
-                onClick={() => router.push('/expenses/add-expense')}
-                className="px-6 py-2 rounded-lg font-medium transition bg-gray-100 text-gray-700 hover:bg-gray-200"
-              >
-                Add Expense
-              </button>
-            </div>
-          </div>
-        </div>
-      </nav>
 
-      <div className="px-6 pb-12">
+      <div className="px-6 pb-12 pt-6">
         <div className="max-w-4xl mx-auto space-y-6">
           <div className="flex justify-between items-center">
             <h1 className="text-3xl font-bold text-gray-800">Dashboard</h1>
